tests/wpt/tests/scroll-animations/scroll-timelines/source-quirks-mode.html

<!-- Quirks mode -->
<html>
<head>
  <title>ScrollTimeline default source in quirks mode</title>
  <link rel="help" href="https://drafts.csswg.org/scroll-animations-1/#dom-scrolltimeline-scrolltimeline">
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <style>
    /* This is just to make it possible for #body1 to be
       "potentially scrollable".

       https://drafts.csswg.org/cssom-view/#potentially-scrollable */
    html {
      overflow: hidden;
    }
  </style>
</head>
<body id=body1></body>
<script>
test(() => {
  try {
    assert_equals(document.scrollingElement.id, 'body1');
    assert_equals(new ScrollTimeline({}).source, body1);

    // Make #body1 "potentially scrollable". This causes the scrollingElement
    // of the document to become null.
    //
    // https://drafts.csswg.org/cssom-view/#dom-document-scrollingelement
    body1.style = 'overflow:scroll';
    assert_equals(new ScrollTimeline({}).source, null);
  } finally {
    body1.style = '';
  }
}, 'Style of <body> is reflected in source attribute in quirks mode');
</script>
</html>
