LayoutTests/imported/w3c/web-platform-tests/scroll-animations/scroll-timelines/source-quirks-mode-expected.txt


FAIL Style of <body> is reflected in source attribute in quirks mode assert_equals: expected null but got Element node <body id="body1" style="overflow: scroll;">
<script>
test...
